Abstract away BackgroundProp
